+++ app/src/main/java/com/example/simplemoneymanager/domain/account/usecases/GetOverallBalanceUseCase.kt
@@ -5,7 +5,7 @@ import com.example.simplemoneymanager.domain.repository.AccountRepository
 
 class GetOverallBalanceUseCase(private val accountRepository: AccountRepository) {
 
-    operator fun invoke(): LiveData<Int> {
+    operator fun invoke(): LiveData<Double> {
         return accountRepository.getOverallBalance()
     }
 }
