+++ app/src/main/java/com/example/simplemoneymanager/presentation/viewModels/AddAccountDialogViewModel.kt
@@ -19,7 +19,7 @@ class AddAccountDialogViewModel(application: Application): AndroidViewModel(appl
 
     private val compositeDisposable = CompositeDisposable()
 
-    fun addAccount(accountName: String, accountBalance: Int, accountColor: String) {
+    fun addAccount(accountName: String, accountBalance: Double, accountColor: String) {
         val account = Account(accountName, accountBalance, accountColor=accountColor)
         val disposable = addAccountUseCase(account).subscribeOn(Schedulers.io())
             .observeOn(AndroidSchedulers.mainThread()).subscribe({
